Fixed warnings found via splint.

+++ bacnet-stack/src/arf.c
@@ -343,7 +343,7 @@ int arf_ack_decode_service_request(
                 decoded_len =
                         decode_octet_string(&apdu[len], len_value_type,
                         &data->fileData[i]);
-                if (decoded_len != len_value_type) {
+                if ((uint32_t)decoded_len != len_value_type) {
                     return -1;
                 }
                 len += decoded_len;
